+++ examples/crab.py
@@ -1,7 +1,19 @@
-from crabpy.client import crab_factory
+# -*- coding: utf-8 -*-
+'''
+This script demonstrates using the crab client directly or through the
+:func:`crabpy.client.crab_request` function.
+'''
+
+from crabpy.client import crab_factory, crab_request
 
 crab = crab_factory()
 
 res = crab.service.ListGemeentenByGewestId(1)
-
 print res
+
+res = crab.service.ListPostkantonsByGemeenteId(71)
+print res
+
+res = crab_request(crab, 'ListGemeentenByGewestId', 1)
+print res
+