+++ examples/crab_proxy.py
@@ -6,7 +6,7 @@ This script show how to connect to the WS-WRAB service through a proxy.
 from crabpy.client import crab_factory
 
 crab = crab_factory(
-    proxy = {
+    proxy={
         'http': 'http://proxy.example.com:3128',
         'https': 'https://httpsproxy.example.com:3128'
     }